@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from functools import wraps
+from functools import WRAPPER_ASSIGNMENTS, wraps
 from itertools import pairwise
 from typing import TYPE_CHECKING, cast
 from warnings import warn
@@ -9,14 +9,14 @@ import h5py
 from packaging.version import Version
 
 from .._core.sparse_dataset import BaseCompressedSparseDataset
-from ..compat import add_note
 
 if TYPE_CHECKING:
-    from collections.abc import Callable
-    from typing import Literal
+    from collections.abc import Callable, Mapping
+    from typing import Any, Literal
 
-    from .._types import StorageType
+    from .._types import ContravariantRWAble, StorageType, _WriteInternal
     from ..compat import H5Group, ZarrGroup
+    from .specs.registry import Writer
 
     Storage = StorageType | BaseCompressedSparseDataset
 
@@ -119,7 +119,8 @@ def check_key(key):
     # elif issubclass(typ, bytes):
     # return key
     else:
-        raise TypeError(f"{key} of type {typ} is an invalid key. Should be str.")
+        msg = f"{key} of type {typ} is an invalid key. Should be str."
+        raise TypeError(msg)
 
 
 # -------------------------------------------------------------------------------
@@ -165,7 +166,7 @@ def _get_display_path(store: Storage) -> str:
     if isinstance(store, BaseCompressedSparseDataset):
         store = store.group
     path = store.name or "??"  # can be None
-    return f'/{path.removeprefix("/")}'
+    return f"/{path.removeprefix('/')}"
 
 
 def add_key_note(
@@ -179,7 +180,7 @@ def add_key_note(
 
     dir = "to" if op == "writ" else "from"
     msg = f"Error raised while {op}ing key {key!r} of {type(store)} {dir} {path}"
-    add_note(e, msg)
+    e.add_note(msg)
 
 
 def report_read_key_on_error(func):
@@ -207,7 +208,8 @@ def report_read_key_on_error(func):
                 store = cast("Storage", arg)
                 break
         else:
-            raise ValueError("No element found in args.")
+            msg = "No element found in args."
+            raise ValueError(msg)
         try:
             return func(*args, **kwargs)
         except Exception as e:
@@ -243,7 +245,8 @@ def report_write_key_on_error(func):
                 store = cast("Storage", arg)
                 break
         else:
-            raise ValueError("No element found in args.")
+            msg = "No element found in args."
+            raise ValueError(msg)
         try:
             return func(*args, **kwargs)
         except Exception as e:
@@ -274,7 +277,8 @@ def _read_legacy_raw(
     if modern_raw:
         if any(k.startswith("raw.") for k in f):
             what = f"File {f.filename}" if hasattr(f, "filename") else "Store"
-            raise ValueError(f"{what} has both legacy and current raw formats.")
+            msg = f"{what} has both legacy and current raw formats."
+            raise ValueError(msg)
         return modern_raw
 
     raw = {}
@@ -285,3 +289,25 @@ def _read_legacy_raw(
     if "varm" in attrs and "raw.varm" in f:
         raw["varm"] = read_attr(f["raw.varm"])
     return raw
+
+
+def zero_dim_array_as_scalar(func: _WriteInternal):
+    """\
+    A decorator for write_elem implementations of arrays where zero-dimensional arrays need special handling.
+    """
+
+    @wraps(func, assigned=WRAPPER_ASSIGNMENTS + ("__defaults__", "__kwdefaults__"))
+    def func_wrapper(
+        f: StorageType,
+        k: str,
+        elem: ContravariantRWAble,
+        *,
+        _writer: Writer,
+        dataset_kwargs: Mapping[str, Any],
+    ):
+        if elem.shape == ():
+            _writer.write_elem(f, k, elem[()], dataset_kwargs=dataset_kwargs)
+        else:
+            func(f, k, elem, _writer=_writer, dataset_kwargs=dataset_kwargs)
+
+    return func_wrapper
