@@ -16,6 +16,7 @@ if TYPE_CHECKING:
     from os import PathLike
     from typing import Literal
 
+    from .._types import ArrayStorageType
     from . import anndata
 
 
@@ -93,7 +94,8 @@ class AnnDataFileManager:
         if filemode is not None:
             self._filemode = filemode
         if self.filename is None:
-            raise ValueError("Cannot open backing file if backing not initialized.")
+            msg = "Cannot open backing file if backing not initialized."
+            raise ValueError(msg)
         self._file = h5py.File(self.filename, self._filemode)
 
     def close(self):
@@ -103,7 +105,7 @@ class AnnDataFileManager:
 
     def _to_memory_mode(self):
         """Close the backing file, forget filename, *do* change to memory mode."""
-        self._adata.__X = self._adata.X[()]
+        self._adata._X = self._adata.X[()]
         self._file.close()
         self._file = None
         self._filename = None
@@ -118,7 +120,7 @@ class AnnDataFileManager:
 
 
 @singledispatch
-def to_memory(x, copy=False):
+def to_memory(x, *, copy: bool = False):
     """Permissivley convert objects to in-memory representation.
 
     If they already are in-memory, (or are just unrecognized) pass a copy through.
@@ -131,27 +133,27 @@ def to_memory(x, copy=False):
 
 @to_memory.register(ZarrArray)
 @to_memory.register(h5py.Dataset)
-def _(x, copy=False):
+def _(x: ArrayStorageType, *, copy: bool = False):
     return x[...]
 
 
 @to_memory.register(BaseCompressedSparseDataset)
-def _(x: BaseCompressedSparseDataset, copy=True):
+def _(x: BaseCompressedSparseDataset, *, copy: bool = False):
     return x.to_memory()
 
 
 @to_memory.register(DaskArray)
-def _(x, copy=False):
+def _(x: DaskArray, *, copy: bool = False):
     return x.compute()
 
 
 @to_memory.register(Mapping)
-def _(x: Mapping, copy=False):
+def _(x: Mapping, *, copy: bool = False):
     return {k: to_memory(v, copy=copy) for k, v in x.items()}
 
 
 @to_memory.register(AwkArray)
-def _(x, copy=False):
+def _(x: AwkArray, *, copy: bool = False):
     from copy import copy as _copy
 
     if copy:
@@ -162,7 +164,8 @@ def _(x, copy=False):
 
 @singledispatch
 def filename(x):
-    raise NotImplementedError(f"Not implemented for {type(x)}")
+    msg = f"Not implemented for {type(x)}"
+    raise NotImplementedError(msg)
 
 
 @filename.register(h5py.Group)
@@ -179,7 +182,8 @@ def _(x):
 
 @singledispatch
 def get_elem_name(x):
-    raise NotImplementedError(f"Not implemented for {type(x)}")
+    msg = f"Not implemented for {type(x)}"
+    raise NotImplementedError(msg)
 
 
 @get_elem_name.register(h5py.Group)
