@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     import numpy as np
     from scipy.sparse import csc_matrix, csr_matrix
 
-    from .compat import Index, SpArray
+    from .compat import CSArray, Index
 
 
 __all__ = ["CSRDataset", "CSCDataset"]
@@ -31,7 +31,7 @@ class _AbstractCSDataset(ABC):
     """Which file type is used on-disk."""
 
     @abstractmethod
-    def __getitem__(self, index: Index) -> float | csr_matrix | csc_matrix | SpArray:
+    def __getitem__(self, index: Index) -> float | csr_matrix | csc_matrix | CSArray:
         """Load a slice or an element from the sparse dataset into memory.
 
         Parameters
@@ -45,7 +45,7 @@ class _AbstractCSDataset(ABC):
         """
 
     @abstractmethod
-    def to_memory(self) -> csr_matrix | csc_matrix | SpArray:
+    def to_memory(self) -> csr_matrix | csc_matrix | CSArray:
         """Load the sparse dataset into memory.
 
         Returns
