@@ -1,15 +1,10 @@
 from __future__ import annotations
 
-import os
-import sys
 from codecs import decode
 from collections.abc import Mapping
-from contextlib import AbstractContextManager
-from dataclasses import dataclass, field
-from functools import singledispatch, wraps
+from functools import partial, singledispatch, wraps
 from importlib.util import find_spec
 from inspect import Parameter, signature
-from pathlib import Path
 from types import EllipsisType
 from typing import TYPE_CHECKING, TypeVar
 from warnings import warn
@@ -18,10 +13,7 @@ import h5py
 import numpy as np
 import pandas as pd
 import scipy
-import scipy.sparse
 from packaging.version import Version
-
-from .exceptiongroups import add_note  # noqa: F401
 
 if TYPE_CHECKING:
     from typing import Any
@@ -31,23 +23,15 @@ if TYPE_CHECKING:
 #############################
 
 
-CAN_USE_SPARSE_ARRAY = Version(scipy.__version__) >= Version("1.11")
-
-if not CAN_USE_SPARSE_ARRAY:
-
-    class SpArray:
-        @staticmethod
-        def __repr__():
-            return "mock scipy.sparse.sparray"
-else:
-    SpArray = scipy.sparse.sparray
+CSMatrix = scipy.sparse.csr_matrix | scipy.sparse.csc_matrix
+CSArray = scipy.sparse.csr_array | scipy.sparse.csc_array
 
 
 class Empty:
     pass
 
 
-Index1D = slice | int | str | np.int64 | np.ndarray
+Index1D = slice | int | str | np.int64 | np.ndarray | pd.Series
 IndexRest = Index1D | EllipsisType
 Index = (
     IndexRest
@@ -56,8 +40,8 @@ Index = (
     | tuple[Index1D, Index1D, EllipsisType]
     | tuple[EllipsisType, Index1D, Index1D]
     | tuple[Index1D, EllipsisType, Index1D]
-    | scipy.sparse.spmatrix
-    | SpArray
+    | CSMatrix
+    | CSArray
 )
 H5Group = h5py.Group
 H5Array = h5py.Dataset
@@ -65,32 +49,16 @@ H5File = h5py.File
 
 
 #############################
-# stdlib
-#############################
-
-
-if sys.version_info >= (3, 11):
-    from contextlib import chdir
-else:
-
-    @dataclass
-    class chdir(AbstractContextManager):
-        path: Path
-        _old_cwd: list[Path] = field(default_factory=list)
-
-        def __enter__(self) -> None:
-            self._old_cwd.append(Path())
-            os.chdir(self.path)
-
-        def __exit__(self, *_exc_info) -> None:
-            os.chdir(self._old_cwd.pop())
-
-
-#############################
 # Optional deps
 #############################
 
 if find_spec("zarr") or TYPE_CHECKING:
+    import zarr
+
+    if Version(zarr.__version__).major > 2:
+        msg = "zarr-python major version > 2 is not supported"
+        raise ImportError(msg)
+
     from zarr.core import Array as ZarrArray
     from zarr.hierarchy import Group as ZarrGroup
 else:
@@ -185,6 +153,16 @@ else:
             return "mock cupy.ndarray"
 
 
+if find_spec("legacy_api_wrap") or TYPE_CHECKING:
+    from legacy_api_wrap import legacy_api  # noqa: TID251
+
+    old_positionals = partial(legacy_api, category=FutureWarning)
+else:
+
+    def old_positionals(*old_positionals):
+        return lambda func: func
+
+
 #############################
 # IO helpers
 #############################
@@ -251,7 +229,7 @@ def _from_fixed_length_strings(value):
 
 
 def _decode_structured_array(
-    arr: np.ndarray, dtype: np.dtype | None = None, copy: bool = False
+    arr: np.ndarray, *, dtype: np.dtype | None = None, copy: bool = False
 ) -> np.ndarray:
     """
     h5py 3.0 now reads all strings as bytes. There is a helper method which can convert these to strings,
