@@ -22,7 +22,12 @@ from anndata._io.specs import (
     get_spec,
 )
 from anndata._io.specs.registry import IORegistryError
-from anndata.compat import CAN_USE_SPARSE_ARRAY, SpArray, ZarrGroup, _read_attr
+from anndata.compat import (
+    CSArray,
+    CSMatrix,
+    ZarrGroup,
+    _read_attr,
+)
 from anndata.experimental import read_elem_as_dask
 from anndata.io import read_elem, write_elem
 from anndata.tests.helpers import (
@@ -66,6 +71,7 @@ def store(request, tmp_path) -> H5Group | ZarrGroup:
 
 sparse_formats = ["csr", "csc"]
 SIZE = 2500
+DEFAULT_SHAPE = (SIZE, SIZE * 2)
 
 
 @pytest.fixture(params=sparse_formats)
@@ -73,15 +79,17 @@ def sparse_format(request):
     return request.param
 
 
-def create_dense_store(store, n_dims: int = 2):
-    X = np.random.randn(*[SIZE * (i + 1) for i in range(n_dims)])
+def create_dense_store(
+    store: str, *, shape: tuple[int, ...] = DEFAULT_SHAPE
+) -> H5Group | ZarrGroup:
+    X = np.random.randn(*shape)
 
     write_elem(store, "X", X)
     return store
 
 
 def create_sparse_store(
-    sparse_format: Literal["csc", "csr"], store: G, shape=(SIZE, SIZE * 2)
+    sparse_format: Literal["csc", "csr"], store: G, shape=DEFAULT_SHAPE
 ) -> G:
     """Returns a store
 
@@ -116,6 +124,7 @@ def create_sparse_store(
 @pytest.mark.parametrize(
     ("value", "encoding_type"),
     [
+        pytest.param(None, "null", id="none"),
         pytest.param("hello world", "string", id="py_str"),
         pytest.param(np.str_("hello world"), "string", id="np_str"),
         pytest.param(np.array([1, 2, 3]), "array", id="np_arr_int"),
@@ -206,6 +215,27 @@ def test_io_spec(store, value, encoding_type):
     assert get_spec(store[key]) == _REGISTRY.get_spec(value)
 
 
+@pytest.mark.parametrize(
+    ("value", "encoding_type"),
+    [
+        pytest.param(np.asarray(1), "numeric-scalar", id="scalar_int"),
+        pytest.param(np.asarray(1.0), "numeric-scalar", id="scalar_float"),
+        pytest.param(np.asarray(True), "numeric-scalar", id="scalar_bool"),  # noqa: FBT003
+        pytest.param(np.asarray("test"), "string", id="scalar_string"),
+    ],
+)
+def test_io_spec_compressed_scalars(store: G, value: np.ndarray, encoding_type: str):
+    key = f"key_for_{encoding_type}"
+    write_elem(
+        store, key, value, dataset_kwargs={"compression": "gzip", "compression_opts": 5}
+    )
+
+    assert encoding_type == _read_attr(store[key].attrs, "encoding-type")
+
+    from_disk = read_elem(store[key])
+    assert_equal(value, from_disk)
+
+
 # Can't instantiate cupy types at the top level, so converting them within the test
 @pytest.mark.gpu
 @pytest.mark.parametrize(
@@ -220,7 +250,7 @@ def test_io_spec(store, value, encoding_type):
 @pytest.mark.parametrize("as_dask", [False, True])
 def test_io_spec_cupy(store, value, encoding_type, as_dask):
     if as_dask:
-        if isinstance(value, sparse.spmatrix):
+        if isinstance(value, CSMatrix):
             value = as_cupy_sparse_dask_array(value, format=encoding_type[:3])
         else:
             value = as_dense_cupy_dask_array(value)
@@ -289,7 +319,7 @@ def test_read_lazy_2d_dask(sparse_format, store):
     ],
 )
 def test_read_lazy_subsets_nd_dask(store, n_dims, chunks):
-    arr_store = create_dense_store(store, n_dims)
+    arr_store = create_dense_store(store, shape=DEFAULT_SHAPE[:n_dims])
     X_dask_from_disk = read_elem_as_dask(arr_store["X"], chunks=chunks)
     X_from_disk = read_elem(arr_store["X"])
     assert_equal(X_from_disk, X_dask_from_disk)
@@ -317,6 +347,14 @@ def test_read_lazy_h5_cluster(sparse_format, tmp_path):
         assert_equal(X_from_disk, X_dask_from_disk)
 
 
+def test_undersized_shape_to_default(store: H5Group | ZarrGroup):
+    shape = (3000, 50)
+    arr_store = create_dense_store(store, shape=shape)
+    X_dask_from_disk = read_elem_as_dask(arr_store["X"])
+    assert (c < s for c, s in zip(X_dask_from_disk.chunksize, shape))
+    assert X_dask_from_disk.shape == shape
+
+
 @pytest.mark.parametrize(
     ("arr_type", "chunks", "expected_chunksize"),
     [
@@ -329,10 +367,10 @@ def test_read_lazy_h5_cluster(sparse_format, tmp_path):
         ("csc", (-1, 10), (SIZE, 10)),
         ("csr", (10, None), (10, SIZE * 2)),
         ("csc", (None, 10), (SIZE, 10)),
-        ("csc", (None, None), (SIZE, SIZE * 2)),
-        ("csr", (None, None), (SIZE, SIZE * 2)),
-        ("csr", (-1, -1), (SIZE, SIZE * 2)),
-        ("csc", (-1, -1), (SIZE, SIZE * 2)),
+        ("csc", (None, None), DEFAULT_SHAPE),
+        ("csr", (None, None), DEFAULT_SHAPE),
+        ("csr", (-1, -1), DEFAULT_SHAPE),
+        ("csc", (-1, -1), DEFAULT_SHAPE),
     ],
 )
 def test_read_lazy_2d_chunk_kwargs(
@@ -594,8 +632,6 @@ def test_read_sparse_array(
     else:
         f = h5py.File(path, "a")
     ad.io.write_elem(f, "mtx", a)
-    if not CAN_USE_SPARSE_ARRAY:
-        pytest.skip("scipy.sparse.cs{r,c}array not available")
     ad.settings.use_sparse_array_on_read = True
     mtx = ad.io.read_elem(f["mtx"])
-    assert issubclass(type(mtx), SpArray)
+    assert issubclass(type(mtx), CSArray)
