@@ -8,15 +8,20 @@ import pandas as pd
 import pytest
 from numba.core.errors import NumbaDeprecationWarning
 
-from anndata import AnnData, ImplicitModificationWarning, read_h5ad, read_loom
+from anndata import AnnData, ImplicitModificationWarning, read_h5ad
+from anndata.io import read_loom
 from anndata.tests.helpers import gen_typed_df_t2_size
-from testing.anndata._helpers import xfail_if_numpy2_loompy
 
-X = np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
-L = np.array([[10, 11, 12], [13, 14, 15], [16, 17, 18]])
+X_ = np.arange(12).reshape((3, 4))
+L = np.arange(12).reshape((3, 4)) + 12
 
 
-def test_creation():
+@pytest.fixture(params=[X_, None])
+def X(request):
+    return request.param
+
+
+def test_creation(X: np.ndarray | None):
     adata = AnnData(X=X, layers=dict(L=L.copy()))
 
     assert list(adata.layers.keys()) == ["L"]
@@ -24,10 +29,11 @@ def test_creation():
     assert "X" not in adata.layers
     assert "some_other_thing" not in adata.layers
     assert (adata.layers["L"] == L).all()
+    assert adata.shape == L.shape
 
 
 def test_views():
-    adata = AnnData(X=X, layers=dict(L=L.copy()))
+    adata = AnnData(X=X_, layers=dict(L=L.copy()))
     adata_view = adata[1:, 1:]
 
     assert adata_view.layers.is_view
@@ -36,13 +42,13 @@ def test_views():
     assert adata_view.layers.keys() == adata.layers.keys()
     assert (adata_view.layers["L"] == adata.layers["L"][1:, 1:]).all()
 
-    adata.layers["S"] = X
+    adata.layers["S"] = X_
 
     assert adata_view.layers.keys() == adata.layers.keys()
     assert (adata_view.layers["S"] == adata.layers["S"][1:, 1:]).all()
 
     with pytest.warns(ImplicitModificationWarning):
-        adata_view.layers["T"] = X[1:, 1:]
+        adata_view.layers["T"] = X_[1:, 1:]
 
     assert not adata_view.layers.is_view
     assert not adata_view.is_view
@@ -51,12 +57,12 @@ def test_views():
 @pytest.mark.parametrize(
     ("df", "homogenous", "dtype"),
     [
-        (lambda: gen_typed_df_t2_size(*X.shape), True, np.object_),
-        (lambda: pd.DataFrame(X**2), False, np.int_),
+        (lambda: gen_typed_df_t2_size(*X_.shape), True, np.object_),
+        (lambda: pd.DataFrame(X_**2), False, np.int_),
     ],
 )
 def test_set_dataframe(homogenous, df, dtype):
-    adata = AnnData(X)
+    adata = AnnData(X_)
     if homogenous:
         with pytest.warns(UserWarning, match=r"Layer 'df'.*dtype object"):
             adata.layers["df"] = df()
@@ -68,7 +74,7 @@ def test_set_dataframe(homogenous, df, dtype):
     assert np.issubdtype(adata.layers["df"].dtype, dtype)
 
 
-def test_readwrite(backing_h5ad):
+def test_readwrite(X: np.ndarray | None, backing_h5ad):
     adata = AnnData(X=X, layers=dict(L=L.copy()))
     adata.write(backing_h5ad)
     adata_read = read_h5ad(backing_h5ad)
@@ -77,11 +83,10 @@ def test_readwrite(backing_h5ad):
     assert (adata.layers["L"] == adata_read.layers["L"]).all()
 
 
-@xfail_if_numpy2_loompy
 @pytest.mark.skipif(find_spec("loompy") is None, reason="loompy not installed")
 def test_readwrite_loom(tmp_path):
     loom_path = tmp_path / "test.loom"
-    adata = AnnData(X=X, layers=dict(L=L.copy()))
+    adata = AnnData(X=X_, layers=dict(L=L.copy()))
 
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=NumbaDeprecationWarning)
@@ -105,7 +110,7 @@ def test_backed():
 
 
 def test_copy():
-    adata = AnnData(X=X, layers=dict(L=L.copy()))
+    adata = AnnData(X=X_, layers=dict(L=L.copy()))
     bdata = adata.copy()
     # check that we don’t create too many references
     assert bdata._layers is bdata.layers._data
@@ -115,13 +120,13 @@ def test_copy():
 
 
 def test_shape_error():
-    adata = AnnData(X=X)
+    adata = AnnData(X=X_)
     with pytest.raises(
         ValueError,
         match=(
             r"Value passed for key 'L' is of incorrect shape\. "
             r"Values of layers must match dimensions \('obs', 'var'\) of parent\. "
-            r"Value had shape \(4, 3\) while it should have had \(3, 3\)\."
+            r"Value had shape \(4, 4\) while it should have had \(3, 4\)\."
         ),
     ):
-        adata.layers["L"] = np.zeros((X.shape[0] + 1, X.shape[1]))
+        adata.layers["L"] = np.zeros((X_.shape[0] + 1, X_.shape[1]))
